+++ services/src/main/java/by/pvt/khudnitsky/payments/services/impl/CardServiceImpl.java
@@ -53,7 +53,7 @@ public class CardServiceImpl extends AbstractService<Card> {
      * @throws SQLException
      */
     @Override
-    public Card getById(int id) throws SQLException {
+    public Card getById(Long id) throws SQLException {
         throw new UnsupportedOperationException();
     }
 
@@ -75,7 +75,7 @@ public class CardServiceImpl extends AbstractService<Card> {
      * @throws SQLException
      */
     @Override
-    public void delete(int id) throws SQLException {
+    public void delete(Long id) throws SQLException {
         throw new UnsupportedOperationException();
     }
 }
